refactor(@dpc-sdp/ripple-tide-api): :recycle: refactor locality script to support budget data
