+++ packages/nuxt-ripple-cli/src/commands/geodata/getBudgetData.ts
@@ -1,0 +1,19 @@
+// import {
+//   vicMapFeatureServerIds,
+//   fetchData,
+//   fetchUsingSimplifiedGeometry,
+//   fetchLGAsMatchingGeometry,
+//   fetchBboxForFeature
+// } from './arcGisApi'
+
+async function getBudgetData(
+  savePath: string,
+  options: {
+    where: string
+    initialOffset: number
+    delay: number
+    perPage: number
+  }
+): Promise<void> {}
+
+export default getBudgetData
