+++ packages/nuxt-ripple-cli/src/commands/index.ts
@@ -4,7 +4,7 @@ import rplMockCommand from './mock'
 import rplInitCommand from './init'
 import rplAddCommand from './add'
 import rplFaviconCommand from './favicon'
-import rplLocalitiesCommand from './localities'
+import rplGeoDataCommand from './geodata'
 
 const program = new commander.Command('ripple')
 
@@ -17,7 +17,7 @@ rippleCli.addCommand(rplMockCommand())
 rippleCli.addCommand(rplInitCommand())
 rippleCli.addCommand(rplAddCommand())
 rippleCli.addCommand(rplFaviconCommand())
-rippleCli.addCommand(rplLocalitiesCommand())
+rippleCli.addCommand(rplGeoDataCommand())
 rippleCli.showHelpAfterError()
 // TODO Add update command for existing sites
